[#167105043] Seat number cannot exceed bus capacity
At the moment, the endpoint allows a booking to be made
if the provided seat number is greater than bus capacity.
This will eventually lead to a situation where there are
more bookings than available seats on the bus. The provided
seat number should have a ceiling value equal to the capacity
of the bus.