[#167123513] Redesign GET /trips API endpoint
At the moment, the endpoint allows a user to filter trips by origin
or destination but, it is not properly designed in that, Swagger
documentation disallow GET request from having a body property.

This redesign will focus on having the origin and destination as
part of the request query. For instance
GET /trips?origin=some_location&destination=some_other_location